requirements.txt
chromadb==1.0.15
sentence-transformers==3.0.1
google-generativeai==0.7.2
python-dotenv==1.0.1
typer==0.16.0
rich==14.0.0




#Login library for debugging--